update deps and pre-commit hook

+++ src/sn_index.py
@@ -8,7 +8,7 @@ def main() -> None:
     output_path = Path("out/sn/index.parquet")
 
     pl.scan_csv(data_path).with_columns(
-        pl.col("year").cast(pl.Int32), pl.col("month").cast(pl.UInt32)
+        pl.col("year").cast(pl.Int32), pl.col("month").cast(pl.Int8)
     ).sink_parquet(output_path)
